Vanilla HTTP client for startscreen.
After some experimentation, it is infeasible
to develop the type of game I want without using
a connection-oriented protocol. I am opting to
use websockets protocol. However, I have done a lot
of work on the client side and would like to catalog
this branch in case I need it as a reference.