Support for basic game client side.
Timer component created when a timer is needed
in-game. Timer will also send WS msg when timer
has finished. GameScreen component added to display
the game to the player. Conditional checks and
message sending added to support basic game. With
this git push, a player in the room can start the
game. One player then gets randomly assigned the
ghoul role and the game begins. Players can move
around the mansion. If the player and ghoul end up
in the same room, the ghoul kills that player and
the player gets a death screen.